Use new conda-forge docker image for compilers
Switches to the `condaforge/linux-anvil-comp7` image for building
packages. This is basically the same as `condaforge/linux-anvil` except
that it does not install the `devtoolset-2` compilers. Meaning it builds
under a non-`root` (regular) user. Also it has access to `yum` for
installing things so can satisfy `yum_requirements.txt`. As both of
these weren't possible on the Anaconda image, several feedstocks were
blocked from migration. This should help unblock them.