Pin `cudnn` globally
The `cudnn` package shipped by `defaults` is batteries included. Though
it currently does not have `run_exports`. To establish some kind of
pinning for the `defaults` package of `cudnn` when used by other
packages, this sets the `cudnn` version to 7. Also it sets the `min_pin`
to respect forward and backward compatibility of the library at the
patch version. Finally as the library maintains backwards compatibility
for each major release, this sets `max_pin` to respect that.

ref: https://docs.nvidia.com/deeplearning/sdk/cudnn-developer-guide/index.html#programming-model