Pin alsa-lib to 1.1.5, the widely-used version.
`alsa-lib` already has a run_exports with max_pin=x.x.x, and up until
two days ago there was only the one version released on conda-forge.
This will set the pin for that version with no required rebuilds. Then
packages can update to the new version of `alsa-lib` with a migrator to
follow.