Pin uhd to 3.15.0.
There are now a few packages that build against `uhd`'s C/C++
library:
    gnuradio-osmosdr
    gnuradio-uhd (gnuradio parent package)
    soapysdr-module-uhd
I'm not aware of any more packages that might depend on `uhd` in the
future, but it would still be nice to have a pin so that keeping these
in sync is easier. `uhd` already has a `run_exports` with
`max_pin=x.x.x`.