Use Quay.io
As DockerHub is implementing consumption limits for downloading images,
which would impact our usage if these are enforced for conda-forge,
this makes the switch to using Quay.io for all of our images. As we push
our images to both container registries, this should be seamless. Though
would require feedstock maintainers to re-render after this is merged to
get this change.

https://www.docker.com/blog/scaling-docker-to-serve-millions-more-developers-network-egress/
https://www.docker.com/blog/docker-hub-image-retention-policy-delayed-and-subscription-updates/
https://www.docker.com/pricing/resource-consumption-updates