Finish Boost 1.78.0 migration
At this point a bit more than 68% of the migration PRs have been merged.
Nearly 27% have open PRs (roughly a quarter of those are passing and the
remainder have varying degrees of work to do). This includes PRs
manually created for the not-solvable or bot-error case (excluding a
couple clearly unmaintained feedstocks). In total this is 95% of the
migration PRs created (most of which have been merged). There are 5%
that either do not need to be migrated, are waiting for parents, need a
version update to be migrated, or have some other issue (like lagging
maintenance) that prevents them from being migrated.

Additionally Boost 1.78.0 has a number of fixes that were needed for
Python 3.11 support. Attempting to backport these to Boost 1.74.0 (the
current Boost version) proved difficult. So this backporting effort has
been abandoned. As there are feedstocks, which require Boost as
part of their Python support, it makes sense to get them onto a newer
Boost compatible with more recent Python versions.

Given all of this, think it is time to close out the Boost 1.78.0
migration and move to using Boost 1.78.0 as the supported Boost in
conda-forge.