Add CUDA 11.8 to sort order in CUDA 12.0 migrator
As CUDA 11.8 is now included by default, conda-smithy will try sort
other CUDA versions relative to 11.8. However the sort order in the CUDA
12.0 migrator didn't include CUDA 11.8 before. As a result, conda-smithy
will error when trying to sort CUDA 11.8. So add CUDA 11.8 to the sort
order in the CUDA 12.0 migrator. This shouldn't effect the CUDA 12.0
migration itself; however, it may mean that users will need to refresh
their copy of the CUDA 12.0 migrator as well.