Match conda-smithy's `--exclusive-file` structure
When using conda-smithy's `--exclusive-file` flag, conda-smithy expects
that the exclusive file will live in a specific location relative to the
`conda_vuild_config.yaml` file used as the exclusive file. In particular
it expects to find the migrations in `./share/conda-forge/migrations`
where `.` contains `conda_vuild_config.yaml`.

However this makes it tricky to test new migration changes in
`conda-forge-pnning` on feedstocks via `--exclusive-file`. As a remedy
create the same directory structure in `conda-forge-pinning` and use a
symlink to map the `migrations` folder to the location `conda-smithy`
expects. This way when using `--exclusive-file` with
`conda-forge-pinning` any local migration changes are picked up and
applied as well.