Close CUDA 12.0 migration
Now that a large majority of feedstocks have been rebuilt or have an
open migration PR, let's go ahead and close out this PR and add CUDA 12
to the matrix.

This adds CUDA 12.0 & CUDA 12.4. Where CUDA 12.0 represents the oldest
CUDA 12 version built against and supported in conda-forge. Similarly
CUDA 12.4 represents the latest CUDA version in conda-forge. This way if
users are still have packages that depend on an older CUDA 12 version,
packages built with CUDA 12.0 will satisfy that need. Also if users are
trying to take advantage of packages built with the latest
optimizations, they can install packages built with the latest CUDA
Toolkit in conda-forge. We keep the latest version pinned to avoid
pulling in partially rolled out CUDA Toolkit updates. Meaning we will
need to go back through and bump this number when a new CUDA Toolkit
becomes available.