Reduce cuda complexity
For years we have struggled in maintainning the multiple cuda versions.

Personally, I feel that Nvidia has done tremendous work in ameliorating
the situation here at Conda-forge.

However, I'm hoping to shed a little of bit of mental weight from cuda
11.8.

IMO the main projects that utilize it:
* pytorch - https://github.com/conda-forge/pytorch-cpu-feedstock/blob/main/recipe/meta.yaml#L58
* tensorflow - https://github.com/conda-forge/tensorflow-feedstock/blob/main/recipe/meta.yaml#L60
* jax - https://github.com/conda-forge/jaxlib-feedstock/blob/main/recipe/meta.yaml#L27
* onnxruntime https://github.com/conda-forge/onnxruntime-feedstock/blob/main/recipe/meta.yaml#L48

have all dropped CUDA 11.8 for lack of maintainer interest.

I think it is time for us to drop it from the global pinnings