Add pin for libsystemd and libudev
This pins the library outputs of the systemd package to version 257,
which is the most recent version that will depend on glibc 2.17. The
next version, 258, dropped support for older kernels and is requiring
the package to depend on glibc 2.34. The ABI of these libraries is
pretty stable, and building against version 257 will still allow users
with glibc 2.34+ to use the version 258 and later packages. A
maintenance branch for version 257 and c_stdlib 2.17 support will be
created for systemd so that this pin can be supported for as long as
needed.

All existing packages have been built against version 257 or earlier,
since the version 258 build has not been merged yet, so no migration is
needed.

See https://github.com/conda-forge/systemd-feedstock/pull/47.